Fixed order of operation in `edit()` method.
The previous order was replacing the new adjustment in the Procedure, then adding the action to the History.
When adding the action to the History, the old version of the adjustment was already gone and the new one was getting saved in both "before" and "after" fields.
When undoing, the variable that was supposed to contain the old version of the Adjustment, contained the new instead and the undoing wasn't actually being done.